Fix for an unstable database state
It happened that one of the switch port had a related interface while
the related interface did not have the port.related_port attribute (which
should not happened but happened) so the code was not able to fix it.
Now it can be fixed. I did not managed to get it work if you want to
remove the existing relation (setting related to None) but now, at least
you can re-add the existing relation (add correct related where it is
missing) and find back a stable state and restart normally from here.