Add customisation to Django Contrib Admin
The default theme of Django Contrib Admin is not really great. This
theme the Django Admin with Re2o logo and colors.

It also comes with a new dynamic-generated menu to navigate in the admin
interface.

A new index page override the global index of Django Contrib Admin to
prepare for new features.

This patch is part of the Aube patchset for Re2o, targeting the use of
Django Contrib Admin as a manage interface for active members.